ADAT receiver/decoder module
this works with the reclocker to provide 8 24-bit parallel outputs from
a serial ADAT input and an 8x oversampled clock; bit sync signal is also
provided for clocking stuff downstream.

There is an ugly generate block to reverse the bits. Some information on
the internet about ADAT uses 0 to indicate the MSB and 23 to indicate the
LSB, defying all common sense. The main shift register should probably
be turned around to eliminate the need for this block.